chown files detected by me
The current need-to-chown detection relies on the user having access
to the root FCCH Public folder, and all intermediate subfolders, all
the way down to the relevant file. If some subfolder has been shared
with a user, they may have been able to create files there without
having access to the whole folder tree, which will cause the code not
to believe that the file is reachable from the root FCCH Public
folder. To counter-act this kind of issue, make each user chown any
file that either they can reach from the root folder, or that I can.